Repurpose sitemap code into something that can be used for iterating over filenames in a directory
