+++ isb_lib/sitemaps/gh_pages_sitemap.py
@@ -1,25 +1,25 @@
+import datetime
 import typing
 from typing import Optional
-
-import requests
+import glob
+import os.path
 
 from isb_lib.sitemaps import SitemapIndexEntry, UrlSetEntry
-from isb_web import isb_solr_query
 
 MAX_URLS_IN_SITEMAP = 50000
 
 
-class ThingUrlSetIterator:
+class GHPagesUrlSetIterator:
     """Iterator class responsible for listing individual urls in an urlset"""
 
     def __init__(
         self,
         sitemap_index: int,
         max_length: int,
-        things: typing.List[typing.Dict[str, str]],
+        filenames: typing.List[str],
     ):
-        self._things: list = things
-        self._thing_index = 0
+        self._filenames: list = filenames
+        self._filename_index = 0
         self._max_length = max_length
         self.num_urls = 0
         self.sitemap_index = sitemap_index
@@ -31,14 +31,15 @@ class ThingUrlSetIterator:
 
     def __next__(self) -> UrlSetEntry:
         # Dont read past the bounds
-        if self._thing_index == len(self._things):
+        if self._filename_index == len(self._filenames):
             raise StopIteration
-        next_thing = self._things[self._thing_index]
-        timestamp_str = next_thing.get("sourceUpdatedTime")
-        next_url_set_entry = UrlSetEntry(next_thing["id"], timestamp_str)
+        next_filename = os.path.splitext(self._filenames[self._filename_index])
+        # TODO: this is a placeholder, need a way to maintain this state somehow
+        timestamp_str = str(datetime.datetime.now())
+        next_url_set_entry = UrlSetEntry(next_filename, timestamp_str)
         # Update the necessary state
         self.num_urls += 1
-        self._thing_index += 1
+        self._filename_index += 1
         self.last_tstamp_str = next_url_set_entry.last_mod_str
         self.last_identifier = next_url_set_entry.identifier
         return next_url_set_entry
@@ -49,43 +50,40 @@ class ThingUrlSetIterator:
         )
 
 
-class ThingSitemapIndexIterator:
+class GHPagesSitemapIndexIterator:
     """Iterator class responsible for listing the individual sitemap files in a sitemap index"""
 
     def __init__(
         self,
-        authority: str = None,
-        num_things_per_file: int = MAX_URLS_IN_SITEMAP,
-        status: int = 200,
-        offset: int = 0,
+        json_directory_path: str,
+        num_things_per_file: int = MAX_URLS_IN_SITEMAP
     ):
         self._last_timestamp_str: Optional[str] = None
         self._last_primary_key: Optional[str] = "0"
-        self._authority = authority
+        self._json_directory_path = json_directory_path
         self._num_things_per_file = num_things_per_file
-        self._status = status
-        self._offset = offset
-        self._last_url_set_iterator: Optional[ThingUrlSetIterator] = None
-        self._rsession = requests.session()
+        self._last_url_set_iterator: Optional[GHPagesUrlSetIterator] = None
         self.num_url_sets = 0
+        self._json_files = glob.glob(os.path.join(json_directory_path, "*.json"))
+        self._json_file_index = 0
 
     def __iter__(self):
         return self
 
-    def __next__(self) -> ThingUrlSetIterator:
+    def __next__(self) -> GHPagesUrlSetIterator:
         if self._last_url_set_iterator is not None:
             # Update our last values with the last ones from the previous iterator
             self._last_timestamp_str = self._last_url_set_iterator.last_tstamp_str
             self._last_primary_key = self._last_url_set_iterator.last_identifier
-        things = isb_solr_query.solr_records_for_sitemap(
-            self._rsession, self._authority, self._offset, self._num_things_per_file
-        )
-        if len(things) == 0:
+
+        if self._json_file_index > len(self._json_files):
             raise StopIteration
-        next_url_set_iterator = ThingUrlSetIterator(
-            self.num_url_sets, self._num_things_per_file, things
+
+        filenames_sublist = self._json_files[self._json_file_index: self._num_things_per_file]
+        next_url_set_iterator = GHPagesUrlSetIterator(
+            self.num_url_sets, self._num_things_per_file, filenames_sublist
         )
         self._last_url_set_iterator = next_url_set_iterator
         self.num_url_sets = self.num_url_sets + 1
-        self._offset = self._offset + self._num_things_per_file
+        self._json_file_index = self._json_file_index + self._num_things_per_file
         return next_url_set_iterator
